requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Mail
Werkzeug
psycopg2-binary
